i love SE

+++ 0x00-hello_world/6-size.c
@@ -14,11 +14,11 @@ int main(void)
 	long long int w;
 	float q;
 
-	printf('Size of a char: %lu byte(s)\n', (unsigned int)sizeof(x));
-	printf('Size of an int: %lu byte(s)\n', (unsigned int)sizeof(z));
-	printf('Size of a long int: %lu byte(s)\n', (unsigned int)sizeof(y));
-	printf('Size of a long long int: %lu byte(s)\n', (unsigned int)sizeof(w));
-	printf('Size of a float: %lu bytes(s)\n', (unsigned int)sizeof(q));
+	printf('Size of a char: %lu byte(s)\n', sizeof(x));
+	printf('Size of an int: %lu byte(s)\n', sizeof(z));
+	printf('Size of a long int: %lu byte(s)\n', sizeof(y));
+	printf('Size of a long long int: %lu byte(s)\n', sizeof(w));
+	printf('Size of a float: %lu bytes(s)\n', sizeof(q));
 	return (0);
 
 }
